Fix duplicate elimination in language selection
Instead of using the Locale.getCountry() string as the stored preference
for the user's language selection, we should probably use
Locale.toString(). For example, this would means storing de_DE instead
of DE.

Otherwise, if the system Locale has the same country as a translation,
but a different language (something like en_DE), the actual German
translation (de_DE) is not shown in the preference list as it is
identified as a duplicate by the comparator used by
LanguagePreference.getData().

Fix it by replacing .getCountry() with .toString() at the relevant
places.